Revert "Revert "removed old files, cleanup, renamed files""
This reverts commit c7d98c040d21e8ade54241edb7420b750ad8b407.
