+++ auth/src/events/listeners/emailChangedListener.ts
@@ -2,7 +2,7 @@ import { Listener, EmailChangedEvent, Subjects } from '@tcosmin/common';
 import { Message } from 'node-nats-streaming';
 import Container from 'typedi';
 import { AuthService } from '../../services/authService';
-import { queueGroupName } from './queue-group-name';
+import { queueGroupName } from './queueGroupName';
 
 export class EmailChangedListener extends Listener<EmailChangedEvent> {
   subject: Subjects.EmailChanged = Subjects.EmailChanged;
