+++ auth/src/events/publishers/accessTokenRevokedPublisher.ts
